requirements.txt
Flask==0.12.2
Flask-Cors==3.0.2
gunicorn==19.0.0
Jinja2==2.9.4
numexpr==2.6.2
numpy==1.12.1
oauthlib==1.0.3
Pillow==4.1.1
requests==2.14.2
requests-file==1.4.1
requests-oauthlib==0.6.1
scipy
tensorflow
keras